pyproject.toml
[project]
name = "brighthive_loadstress_cdk"
version = "0.1.0"
description = "AWS CDK infrastructure for BrightAgent load/stress scenarios at production scale"
authors = [
    { name = "Hikuri Chinca", email = "kuri@brighthive.io" }
]
readme = "README.md"
requires-python = ">=3.11"
keywords = []
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
]

# ============================================================================
# PRODUCTION DEPENDENCIES (Required for CDK deployment)
# ============================================================================
dependencies = [
    # AWS CDK Infrastructure
    "aws-cdk-lib>=2.139.1",
    "constructs>=10.0.0",

    # AWS SDK
    "boto3>=1.34.0",

    # Configuration & Settings
    "pyyaml>=6.0.1",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",

    # HTTP & Async I/O
    "httpx>=0.27.0",
    "aiofiles>=24.1.0",

    # Logging
    "structlog>=24.4.0",
]

# ============================================================================
# OPTIONAL DEPENDENCIES (Install with: uv pip install -e ".[dev]")
# ============================================================================
[project.optional-dependencies]

# Local development (linting, formatting, type checking)
dev = [
    # Testing framework
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",

    # Code quality
    "ruff>=0.8.0",              # Fast linter & formatter
    "mypy>=1.13.0",             # Type checking
    "pre-commit>=4.0.0",        # Git hooks

    # Interactive development
    "ipython>=8.29.0",          # Better REPL
    "ipdb>=0.13.13",            # Debugger
]

# PySpark local testing (test data generation scripts WITHOUT EMR cluster)
pyspark-testing = [
    "pyspark>=3.5.0",           # Local Spark for unit tests
    "faker>=22.0.0",            # Realistic mock data generation
    "pytest>=8.3.0",            # Testing framework
]

# Advanced load testing & profiling
loadstress-advanced = [
    "pytest-benchmark>=5.1.0",  # Performance benchmarking
    "memory-profiler>=0.61.0",  # Memory usage analysis
    "line-profiler>=4.2.0",     # Line-by-line profiling
]

# Security scanning
security = [
    "safety>=3.0.0",            # Check for known vulnerabilities
    "pip-audit>=2.7.0",         # Audit dependencies
    "bandit>=1.7.0",            # Security linter
]

# Observability & tracing
observability = [
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
]

# Everything for local development
all-dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-benchmark>=5.1.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    "ipython>=8.29.0",
    "ipdb>=0.13.13",
    "pyspark>=3.5.0",
    "faker>=22.0.0",
    "memory-profiler>=0.61.0",
    "line-profiler>=4.2.0",
    "safety>=3.0.0",
    "pip-audit>=2.7.0",
    "bandit>=1.7.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/brighthive_loadstress_cdk"]

# ============================================================================
# UV TOOL CONFIGURATION
# ============================================================================
[tool.uv]
# Default dev dependencies when running: uv sync
dev-dependencies = [
    # Core testing & quality
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",

    # PySpark for local testing (NEW!)
    "pyspark>=3.5.0",
    "faker>=22.0.0",

    # Interactive development
    "ipython>=8.29.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=brighthive_loadstress_cdk",
    "--cov-report=term-missing",
    "--cov-report=html",
]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/.venv/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "@abstractmethod",
]

[tool.bandit]
exclude_dirs = ["tests", ".venv", "venv", "build", "dist"]
skips = ["B101"]  # allow assert in tests
tests = ["B201", "B301", "B302", "B303", "B304", "B305", "B306", "B307", "B308", "B310", "B311", "B312", "B313", "B314", "B315", "B316", "B317", "B318", "B319", "B320", "B321", "B323", "B324", "B325", "B401", "B402", "B403", "B404", "B405", "B406", "B407", "B408", "B409", "B410", "B411", "B412", "B413", "B501", "B502", "B503", "B504", "B505", "B506", "B507", "B508", "B509", "B601", "B602", "B603", "B604", "B605", "B606", "B607", "B608", "B609", "B610", "B611", "B612", "B701", "B702", "B703"]
